requirements.txt
django==4.1.6
django-extensions
django-filter
requests
transformers==4.26.1
torch==1.13.1
djangorestframework